--- a/src/coreclr/src/tools/Common/Sorting/ICompareAsEqualAction.cs
+++ b/src/coreclr/src/tools/Common/Sorting/ICompareAsEqualAction.cs
@@ -1,0 +1,28 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics;
+
+namespace ILCompiler
+{
+    internal interface ICompareAsEqualAction
+    {
+        void CompareAsEqual();
+    }
+
+    internal struct RequireTotalOrderAssert : ICompareAsEqualAction
+    {
+        public void CompareAsEqual()
+        {
+            Debug.Assert(false);
+        }
+    }
+
+    internal struct AllowDuplicates : ICompareAsEqualAction
+    {
+        public void CompareAsEqual()
+        {
+        }
+    }
+}