--- a/src/coreclr/src/tools/Common/Sorting/ISortableDataStructureAccessor.cs
+++ b/src/coreclr/src/tools/Common/Sorting/ISortableDataStructureAccessor.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace ILCompiler
+{
+    internal interface ISortableDataStructureAccessor<T, TDataStructure>
+    {
+        T GetElement(TDataStructure dataStructure, int i);
+        void SetElement(TDataStructure dataStructure, int i, T value);
+        void SwapElements(TDataStructure dataStructure, int i, int i2);
+        void Copy(TDataStructure source, int sourceIndex, T[] target, int destIndex, int length);
+        void Copy(T[] source, int sourceIndex, TDataStructure target, int destIndex, int length);
+        int GetLength(TDataStructure dataStructure);
+    }
+}