--- a/src/coreclr/src/tools/crossgen2/ILCompiler.DependencyAnalysisFramework/DependencyAnalyzer.cs
+++ b/src/coreclr/src/tools/crossgen2/ILCompiler.DependencyAnalysisFramework/DependencyAnalyzer.cs
@@ -311,7 +311,7 @@ namespace ILCompiler.DependencyAnalysisFramework
                 } while (_markStack.Count != 0);
 
                 if (_resultSorter != null)
-                    _markedNodes.ParallelSortAllowDuplicates(_resultSorter);
+                    _markedNodes.MergeSortAllowDuplicates(_resultSorter);
 
                 _markedNodesFinal = _markedNodes.ToImmutableArray();
                 _markedNodes = null;