@davidwrighton 2020-05-06
@nattress The extra temporary allocations from this algorithm are about the same size as the initial data structure to sort. In the context of crossgen2, which the graph is often quite large, the sortable structure itself is a small fraction of the size, and the size cost is really small. I didn't see a memory size regression of significance, although the test cases you've been looking at are much bigger, and may show a more significant regression.

@trylek The reason the comparers are slow is that the objects they are comparing are spread all over the GC heap, with extremely poor locality. Its not really a matter of performing any individual comparison faster, its more about reducing the latency in reading the interesting bits of memory.