@MichalStrehovsky 2020-05-06
> Using standard sorts 121.6 seconds
> Using parallel merge sort 96.9 seconds

This is a pretty surprising bit of information. I won't have time to look into this, but as a datapoint - the CoreRT AOT compiler is architected pretty similarly (compilation is multithreaded and the sorting of everything happens on a single thread).

In the CoreRT compiler the sorting contributes to only ~2% of wallclock time (I did a quick measurement for "CoreRT compiler compiling CoreRT compiler" because that's what I was playing with last weekend - it's a bigger app than a Hello world, but of course smaller than many others).

I'm kind of wondering why ReadyToRun needs to sort so much more and whether the main problem is that we're generating too many things.