@davidwrighton 2020-05-06
@MichalStrehovsky yes, it is clear that the expansion rate of R2R images is greater than it should be, and that is the real problem. I'm looking into that as well, as it leads to more extreme problems like oversize files, and such. However, this fix is nice as it substantially reduces the cycle time of looking at changes. I suspect that what is happening is that crossgen2 is getting TLB cache thrashing, which can be disastrous for performance, and this fix is spreading that load across various different cores thus leading to significantly better perf.