@vectorix 2020-05-07
This method doesn't look right.

         public void SwapElements(T[] dataStructure, int i, int i2)
        {
            T temp = dataStructure[i];
            dataStructure[i] = dataStructure[i + 1];
            dataStructure[i + 1] = temp;
        }

It doesn't swap i with i2, it swaps i with i +1 and ignores i2.

Since the caller passes i + 1 in i2 this parameter may be redundant, but then the name should probably be something else, e.g. SwapWithNext.