@davidwrighton 2020-05-07
@vectorix wow, that's an interesting copy/paste bug. Fortunately, the only use of this function is to swap with next, so the behavior is … correct in the product, but a function that wrong should be fixed. Thank you for your review.